+++ src/main/java/com/github/sweintritt/jmus/State.java
@@ -1,7 +1,0 @@
-package com.github.sweintritt.jmus;
-
-public enum State {
-    PLAYING,
-    STOPPED,
-    SEARCHING
-}